fix: 修复日志聚类初始化分组不生效问题 # Reviewed, transaction id: 17982
