+++ bklog/web/src/views/retrieve/result-table-panel/log-clustering/components/quick-open-cluster-step/quick-open-cluster.tsx
@@ -30,9 +30,9 @@ import { Component as tsc } from 'vue-tsx-support';
 import { From } from 'bk-magic-vue';
 
 import $http from '../../../../../../api';
-import { formatDate, deepClone } from '../../../../../../common/util';
-import { handleTransformToTimestamp } from '../../../../../../components/time-range/utils';
+import { deepClone } from '../../../../../../common/util';
 import clusterImg from '../../../../../../images/cluster-img/cluster.png';
+import FilterRule from './filter-rule';
 
 import './quick-open-cluster.scss';
 
@@ -48,20 +48,9 @@ export default class QuickOpenCluster extends tsc<IProps> {
   @Prop({ type: Array, required: true }) datePickerValue: Array<any>;
   @Prop({ type: Object, required: true }) retrieveParams: object;
   @Ref('quickClusterFrom') quickClusterFromRef: From;
+  @Ref('filterRule') filterRuleRef;
 
   isShowDialog = false;
-  conditionList = [
-    // 过滤条件对比
-    { id: '=', name: '=' },
-    { id: '!=', name: '!=' },
-    { id: 'LIKE', name: 'LIKE' },
-    { id: 'NOT LIKE', name: 'NOT LIKE' },
-  ];
-  comparedList = [
-    { id: 'and', name: 'AND' },
-    { id: 'or', name: 'OR' },
-  ];
-  operateIndex = 0;
   formData = {
     clustering_fields: '',
     filter_rules: [],
@@ -76,34 +65,8 @@ export default class QuickOpenCluster extends tsc<IProps> {
         trigger: 'blur',
       },
     ],
-    filter_rules: [
-      {
-        validator: this.checkFilterRules,
-        trigger: 'blur',
-      },
-    ],
   };
   popoverInstance = null;
-
-  get isShowAddFilterIcon() {
-    const rules = this.formData.filter_rules;
-    if (
-      !rules.length ||
-      (rules.slice(-1)[0].fields_name !== '' && rules.length === 1) ||
-      rules.slice(-1)[0].value.length > 0
-    )
-      return true;
-    return false;
-  }
-
-  get filterSelectList() {
-    return this.totalFields
-      .filter(item => !/^__dist/.test(item.field_name) && item.field_type !== '__virtual__')
-      .map(el => {
-        const { field_name: id, field_alias: alias } = el;
-        return { id, name: alias ? `${id}(${alias})` : id };
-      });
-  }
 
   get clusterField() {
     return this.totalFields
@@ -130,7 +93,9 @@ export default class QuickOpenCluster extends tsc<IProps> {
   handleAccessCluster() {
     this.isShowDialog = true;
   }
-  handleConfirmSubmit() {
+  async handleConfirmSubmit() {
+    const isRulePass = await this.filterRuleRef.handleCheckRuleValidate();
+    if (!isRulePass) return;
     this.quickClusterFromRef.validate().then(async () => {
       try {
         const data = {
@@ -227,71 +192,6 @@ export default class QuickOpenCluster extends tsc<IProps> {
       this.formData = this.cloneFormData;
     }
   }
-  handleFieldChange(fieldName: string, index: number) {
-    const field = this.totalFields.find(item => item.field_name === fieldName) ?? {};
-    Object.assign(this.formData.filter_rules[index], {
-      ...field,
-      value: [],
-    });
-    const requestFields = this.fieldsKeyStrList();
-    this.queryValueList(requestFields);
-  }
-  async queryValueList(fields = []) {
-    if (!fields.length) return;
-    const tempList = handleTransformToTimestamp(this.datePickerValue);
-    try {
-      const res = await $http.request('retrieve/getAggsTerms', {
-        params: {
-          index_set_id: this.$route.params.indexId,
-        },
-        data: {
-          keyword: this.retrieveParams?.keyword ?? '*',
-          fields,
-          start_time: formatDate(tempList[0] * 1000),
-          end_time: formatDate(tempList[1] * 1000),
-        },
-      });
-      this.formData.filter_rules.forEach(item => {
-        item.valueList =
-          res.data.aggs_items[item.fields_name]?.map(item => ({
-            id: item.toString(),
-            name: item.toString(),
-          })) ?? [];
-      });
-    } catch (err) {
-      this.formData.filter_rules.forEach(item => (item.valueList = []));
-    }
-  }
-  fieldsKeyStrList() {
-    const fieldsStrList = this.formData.filter_rules
-      .filter(item => item.field_type !== 'text' && item.es_doc_values)
-      .map(item => item.fields_name);
-    return Array.from(new Set(fieldsStrList));
-  }
-  handleValueBlur(operateItem, val: string) {
-    if (!operateItem.value.length && !!val) operateItem.value.push(val);
-  }
-  checkFilterRules() {
-    return this.formData.filter_rules.every(item => !!item.value.length && item.fields_name);
-    // return !!item.value.length && item.fields_name;
-  }
-  handleDeleteSelect(index: number) {
-    this.formData.filter_rules.splice(index, 1);
-    (this.$refs[`fieldSelectRef-${index}`] as any).close();
-  }
-  handleAddFilterRule() {
-    this.formData.filter_rules.push({
-      fields_name: '', // 过滤规则字段名
-      op: '=', // 过滤规则操作符号
-      value: [], // 过滤规则字段值
-      logic_operator: 'and',
-      valueList: [],
-    });
-    this.$nextTick(() => {
-      const index = this.formData.filter_rules.length - 1;
-      (this.$refs[`fieldSelectRef-${index}`] as any).show();
-    });
-  }
   render() {
     const accessDialogSlot = () => (
       <bk-dialog
@@ -353,89 +253,13 @@ export default class QuickOpenCluster extends tsc<IProps> {
             label={$i18n.t('过滤规则')}
             property='filter_rules'
           >
-            <div class='filter-rule'>
-              {this.formData.filter_rules.map((item, index) => (
-                <div class='filter-rule filter-rule-item'>
-                  {!!this.formData.filter_rules.length && !!index && !!item.fields_name && (
-                    <bk-select
-                      class='icon-box and-or mr-neg1'
-                      v-model={item.logic_operator}
-                      clearable={false}
-                    >
-                      {this.comparedList.map(option => (
-                        <bk-option
-                          id={option.id}
-                          name={option.name}
-                        ></bk-option>
-                      ))}
-                    </bk-select>
-                  )}
-                  <bk-select
-                    ref={`fieldSelectRef-${index}`}
-                    class={['min-100 mr-neg1 above', { 'is-not-error': !!item.fields_name }]}
-                    v-model={item.fields_name}
-                    clearable={false}
-                    popover-min-width={150}
-                    searchable
-                    on-selected={fieldsName => this.handleFieldChange(fieldsName, index)}
-                  >
-                    {this.filterSelectList.map(option => (
-                      <bk-option
-                        id={option.id}
-                        name={option.name}
-                      ></bk-option>
-                    ))}
-                    <div
-                      style='cursor: pointer'
-                      slot='extension'
-                      onClick={() => this.handleDeleteSelect(index)}
-                    >
-                      <i class='bk-icon icon-close-circle'></i>
-                      <span style='margin-left: 4px;'>{$i18n.t('删除')}</span>
-                    </div>
-                  </bk-select>
-                  {!!item.fields_name && (
-                    <bk-select
-                      class='icon-box mr-neg1 condition'
-                      v-model={item.op}
-                      clearable={false}
-                      popover-min-width={100}
-                    >
-                      {this.conditionList.map(option => (
-                        <bk-option
-                          id={option.id}
-                          name={option.name}
-                        ></bk-option>
-                      ))}
-                    </bk-select>
-                  )}
-                  {!!item.fields_name && (
-                    <div onClick={() => (this.operateIndex = index)}>
-                      <bk-tag-input
-                        class={['mr-neg1 min-100 above', { 'is-not-error': !!item.value.length }]}
-                        v-model={item.value}
-                        content-width={232}
-                        list={item.valueList}
-                        max-data={1}
-                        placeholder={$i18n.t('请输入')}
-                        trigger='focus'
-                        allow-auto-match
-                        allow-create
-                        on-blur={v => this.handleValueBlur(item, v)}
-                      ></bk-tag-input>
-                    </div>
-                  )}
-                </div>
-              ))}
-              {this.isShowAddFilterIcon && (
-                <button
-                  class='icon-box'
-                  onClick={this.handleAddFilterRule}
-                >
-                  <i class='bk-icon icon-plus-line'></i>
-                </button>
-              )}
-            </div>
+            <FilterRule
+              ref='filterRule'
+              v-model={this.formData.filter_rules}
+              date-picker-value={this.datePickerValue}
+              retrieve-params={this.retrieveParams}
+              total-fields={this.totalFields}
+            ></FilterRule>
           </bk-form-item>
           <bk-form-item
             label={$i18n.t('告警配置')}
