+++ bklog/web/src/views/retrieve/result-table-panel/log-clustering/index.vue
@@ -608,7 +608,7 @@
         const isActiveCluster = await this.confirmClusterStepStatus();
         if (isActiveCluster) {
           this.filterGroupList();
-          this.initTableOperator(); // 初始化分组下拉列表
+          await this.initTableOperator(); // 初始化分组下拉列表
           this.requestFinger();
           this.stopPolling();
         }
