+++ bklog/web/src/views/retrieve/setting-modal/log-cluster.vue
@@ -126,101 +126,14 @@
         </bk-form-item>
         <!-- 过滤规则 -->
         <div style="margin-bottom: 40px">
-          <p style="height: 32px">{{ $t('过滤规则') }}</p>
-          <div class="filter-rule">
-            <div
-              v-for="(item, index) of formData.filter_rules"
-              class="filter-rule filter-rule-item"
-              :key="index"
-            >
-              <bk-select
-                v-if="formData.filter_rules.length !== 0 && index !== 0 && item.fields_name !== ''"
-                class="icon-box and-or mr-neg1"
-                v-model="item.logic_operator"
-                :clearable="false"
-                :disabled="!globalEditable"
-              >
-                <bk-option
-                  v-for="option in comparedList"
-                  :id="option.id"
-                  :key="option.id"
-                  :name="option.name"
-                >
-                </bk-option>
-              </bk-select>
-
-              <bk-select
-                v-if="!isCloseSelect"
-                v-model="item.fields_name"
-                :class="['min-100 mr-neg1 above', item.fields_name === '' && isFieldsError ? 'rule-error' : '']"
-                :clearable="false"
-                :disabled="!globalEditable"
-                :popover-min-width="150"
-                searchable
-                @blur="blurFilter"
-                @selected="fieldsName => handleFieldChange(fieldsName, index)"
-              >
-                <bk-option
-                  v-for="option in filterSelectList"
-                  :id="option.id"
-                  :key="option.id"
-                  :name="option.name"
-                >
-                </bk-option>
-                <template #extension>
-                  <div
-                    style="cursor: pointer"
-                    @click="handleDeleteSelect(index)"
-                  >
-                    <i class="bk-icon icon-close-circle"></i>{{ $t('删除') }}
-                  </div>
-                </template>
-              </bk-select>
-
-              <bk-select
-                v-if="item.fields_name !== ''"
-                style="color: #3a84ff"
-                class="icon-box mr-neg1"
-                v-model="item.op"
-                :clearable="false"
-                :disabled="!globalEditable"
-                :popover-min-width="100"
-              >
-                <bk-option
-                  v-for="option in conditionList"
-                  :id="option.id"
-                  :key="option.id"
-                  :name="option.name"
-                >
-                </bk-option>
-              </bk-select>
-
-              <div @click="handleInputTag(index)">
-                <bk-tag-input
-                  v-if="item.fields_name !== ''"
-                  v-model="item.value"
-                  :class="['mr-neg1 min-100 above', !item.value.length && isFilterRuleError ? 'rule-error' : '']"
-                  :content-width="232"
-                  :list="item.valueList"
-                  :max-data="1"
-                  :placeholder="$t('请输入')"
-                  trigger="focus"
-                  allow-auto-match
-                  allow-create
-                  @blur="handleValueBlur"
-                >
-                </bk-tag-input>
-              </div>
-            </div>
-            <button
-              v-if="isShowAddFilterIcon"
-              class="icon-box"
-              :disabled="!globalEditable"
-              @click="addFilterRule"
-            >
-              <i class="bk-icon icon-plus-line"></i>
-            </button>
-          </div>
+          <p style="height: 24px; font-size: 12px">{{ $t('过滤规则') }}</p>
+          <FilterRule
+            ref="filterRuleRef"
+            v-model="formData.filter_rules"
+            :total-fields="totalFields"
+            :date-picker-value="datePickerValue"
+            :retrieve-params="retrieveParams"
+          ></FilterRule>
         </div>
         <!-- 聚类规则 -->
         <rule-table
@@ -279,14 +192,13 @@
 </template>
 
 <script>
-  import { formatDate } from '@/common/util';
-
-  import { handleTransformToTimestamp } from '../../../components/time-range/utils';
   import RuleTable from './rule-table';
+  import FilterRule from '../result-table-panel/log-clustering/components/quick-open-cluster-step/filter-rule';
 
   export default {
     components: {
       RuleTable,
+      FilterRule,
     },
     props: {
       globalEditable: {
@@ -327,7 +239,6 @@
         isShowAddFilterIcon: true, // 是否显示过滤规则增加按钮
         isShowSubmitDialog: false, // 是否展开保存弹窗
         isHandle: false, // 保存loading
-        filterSelectList: [], // 过滤条件选项
         isFilterRuleError: false, // 过滤规则未填警告
         isFieldsError: false, // 未选过滤条件字段警告
         isCloseSelect: false, // 过滤规则下拉框隐藏
@@ -354,39 +265,9 @@
           filter_rules: [], // 过滤规则
           signature_enable: false,
         },
-        conditionList: [
-          // 过滤条件对比
-          { id: '=', name: '=' },
-          { id: '!=', name: '!=' },
-          { id: 'LIKE', name: 'LIKE' },
-          { id: 'NOT LIKE', name: 'NOT LIKE' },
-        ],
-        comparedList: [
-          { id: 'and', name: 'AND' },
-          { id: 'or', name: 'OR' },
-        ],
-        operateIndex: 0, // 赋值过滤字段的操作的当前下标
         isShowFingerTips: false,
         isActive: false,
       };
-    },
-    watch: {
-      'formData.filter_rules': {
-        deep: true,
-        handler(val) {
-          if (val.length === 0) {
-            this.isShowAddFilterIcon = true;
-            return;
-          }
-          if ((val.slice(-1)[0].fields_name !== '' && val.length === 1) || val.slice(-1)[0].value.length > 0) {
-            this.isShowAddFilterIcon = true;
-          }
-          if (val.slice(-1)[0].fields_name === '') {
-            this.isShowAddFilterIcon = false;
-          }
-          this.isFilterRuleError = false;
-        },
-      },
     },
     mounted() {
       this.initList();
@@ -429,10 +310,6 @@
           // 当前回填的字段如果在聚类字段列表里找不到则赋值为空需要用户重新赋值
           const isHaveFieldsItem = this.clusterField.find(item => item.id === res.data.clustering_fields);
           if (!isHaveFieldsItem) this.formData.clustering_fields = '';
-          this.$nextTick(() => {
-            const requestFields = this.fieldsKeyStrList();
-            this.queryValueList(requestFields);
-          });
         } catch (e) {
           console.warn(e);
         } finally {
@@ -455,17 +332,8 @@
             const { field_name: id, field_alias: alias } = el;
             return { id, name: alias ? `${id}(${alias})` : id };
           });
-        this.filterSelectList = this.totalFields
-          .filter(item => !/^__dist/.test(item.field_name) && item.field_type !== '__virtual__')
-          .map(el => {
-            const { field_name: id, field_alias: alias } = el;
-            return { id, name: alias ? `${id}(${alias})` : id };
-          });
         // 日志聚类且数据指纹同时打开则不请求默认值
         this.requestCluster(false);
-        // if (isActive) {
-        //   this.requestCluster(false);
-        // }
       },
       /**
        * @desc: 数据指纹开关
@@ -499,23 +367,9 @@
           this.requestCluster(true);
         }
       },
-      addFilterRule() {
-        this.formData.filter_rules.push({
-          fields_name: '', // 过滤规则字段名
-          op: '=', // 过滤规则操作符号
-          value: [], // 过滤规则字段值
-          logic_operator: 'and',
-          valueList: [],
-        });
-      },
-      blurFilter() {
-        if (this.formData.filter_rules?.length > 0) {
-          this.isFilterRuleError = this.formData.filter_rules.some(el => !el.value.length);
-          this.isFieldsError = this.formData.filter_rules.some(el => el.fields_name === '');
-        }
-      },
-      handleSubmit() {
-        this.blurFilter();
+      async handleSubmit() {
+        const isRulePass = await this.$refs.filterRuleRef.handleCheckRuleValidate();
+        if (!isRulePass) return;
         this.$refs.validateForm.validate().then(
           () => {
             if (this.isFilterRuleError || this.isFieldsError) return;
@@ -571,69 +425,6 @@
           () => {},
         );
       },
-      // 字段改变
-      handleFieldChange(fieldName, index) {
-        const field = this.totalFields.find(item => item.field_name === fieldName) ?? {};
-        Object.assign(this.formData.filter_rules[index], {
-          ...field,
-          value: [],
-        });
-        const requestFields = this.fieldsKeyStrList();
-        this.queryValueList(requestFields);
-      },
-      async queryValueList(fields = []) {
-        if (!fields.length) return;
-        const tempList = handleTransformToTimestamp(this.datePickerValue);
-        try {
-          const res = await this.$http.request('retrieve/getAggsTerms', {
-            params: {
-              index_set_id: this.$route.params.indexId,
-            },
-            data: {
-              keyword: this.retrieveParams?.keyword ?? '*',
-              fields,
-              start_time: formatDate(tempList[0] * 1000),
-              end_time: formatDate(tempList[1] * 1000),
-            },
-          });
-          this.formData.filter_rules.forEach(item => {
-            item.valueList =
-              res.data.aggs_items[item.fields_name]?.map(item => ({
-                id: item.toString(),
-                name: item.toString(),
-              })) ?? [];
-          });
-        } catch (err) {
-          this.formData.filter_rules.forEach(item => (item.valueList = []));
-        }
-      },
-      fieldsKeyStrList() {
-        const fieldsStrList = this.formData.filter_rules
-          .filter(item => item.field_type !== 'text' && item.es_doc_values)
-          .map(item => item.fields_name);
-        return Array.from(new Set(fieldsStrList));
-      },
-      /**
-       * @desc: 赋值过滤字段的下标
-       * @param { Number } index 下标
-       */
-      handleInputTag(index) {
-        this.operateIndex = index;
-      },
-      handleValueBlur(val) {
-        const operateItem = this.formData.filter_rules[this.operateIndex];
-        if (!operateItem.value.length && val !== '') {
-          operateItem.value.push(val);
-        }
-      },
-      handleDeleteSelect(index) {
-        this.formData.filter_rules.splice(index, 1);
-        this.isCloseSelect = true;
-        // 删除非最后一条过滤规则时隐藏下拉框
-        this.$nextTick(() => {
-          this.isCloseSelect = false;
-        });
-      },
       resetPage() {
         this.$emit('reset-page');
       },
@@ -654,6 +445,14 @@
       margin-top: 16px;
     }
 
+    :deep(.tooltips-icon) {
+      /* stylelint-disable-next-line declaration-no-important */
+      top: -24px !important;
+
+      /* stylelint-disable-next-line declaration-no-important */
+      right: 818px !important;
+    }
+
     .setting-item {
       display: flex;
       align-items: center;
@@ -663,79 +462,6 @@
         margin-left: 8px;
         font-size: 18px;
         color: #979ba5;
-      }
-    }
-
-    .filter-rule {
-      display: flex;
-      flex-wrap: wrap;
-
-      .icon-box {
-        min-width: 32px;
-        height: 32px;
-        font-size: 14px;
-        line-height: 28px;
-        text-align: center;
-        cursor: pointer;
-        background: #fff;
-        border: 1px solid #c4c6cc;
-
-        :deep(.bk-select-name) {
-          /* stylelint-disable-next-line declaration-no-important */
-          padding: 0 !important;
-        }
-
-        .icon-plus-line {
-          color: #3a84ff;
-        }
-      }
-    }
-
-    .filter-rule-item {
-      margin-bottom: 6px;
-
-      :deep(.bk-select-angle) {
-        display: none;
-      }
-
-      :deep(.bk-select) {
-        border-radius: 0;
-      }
-
-      :deep(.bk-form-control) {
-        width: 140px;
-        border-radius: 0;
-      }
-
-      .and-or {
-        min-width: 62px;
-        font-size: 12px;
-        color: #ff9c01;
-      }
-
-      .min-100 {
-        min-width: 100px;
-        max-height: 32px;
-      }
-
-      .mr-neg1 {
-        position: relative;
-        margin-right: -1px;
-      }
-
-      .above {
-        z-index: 99;
-      }
-    }
-
-    .rule-error {
-      :deep(.bk-tag-input) {
-        border-color: #ff5656;
-      }
-
-      &.bk-select {
-        /* stylelint-disable-next-line declaration-no-important */
-        border-color: #ff5656 !important;
       }
     }
